feature: moving everyone to a `std::breakpoint_if_debugging()` style approach, where `DEBUG_BREAK_H()` does not cause the running application to b0rk when NO debugger is attached.
Of course there are issues: only two of the major platforms provide a reasonable/known way to detect the presence of an actual debugger, so we have to resort to "heuristics" shenanigans for the others: we will assume any DEBUG BUILD (`#define _DEBUG`) will expect a debugger and thus produce an active `debug_break()`, while any non-debug, i.e _release_ build is assumed to NOT have a debugger attached.

TODO: provide a simple user-definable override to kill this mess: the application code writer knows best, in the end.  ;-)